+++ lib/gamtools/tests/test_parser.py
@@ -14,9 +14,7 @@ def test_fixed_threshold():
 
     fitting_result = args.fitting_function(np.arange(100))
 
-    assert fitting_result['read_threshold'] == 4
-    assert 'counts' in fitting_result
-    assert 'breaks' in fitting_result
+    assert fitting_result['coverage_threshold'] == 4
     assert fitting_result['params'] is None
 
 
@@ -26,9 +24,15 @@ def test_fixed_threshold_5():
 
     fitting_result = args.fitting_function(np.arange(100))
 
-    assert fitting_result['read_threshold'] == 5
-    assert 'counts' in fitting_result
-    assert 'breaks' in fitting_result
+    assert fitting_result['coverage_threshold'] == 5.0
     assert fitting_result['params'] is None
 
 
+def test_fixed_threshold_process_nps():
+
+    args = main.parser.parse_args(['process_nps', '-g', '/dev/null', '-x', '4', '/dev/null'])
+
+    fitting_result = args.fitting_function(np.arange(100))
+
+    assert fitting_result['coverage_threshold'] == 4
+    assert fitting_result['params'] is None