+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow the doit dependencies database to be reset
 - Calculate coverage on individual bamfiles instead of all at once
 - Calculate intrinsic bias of a contact matrix
+- Added functions for calculating normalized pointwise mutual information (NPMI)
+- Added a CONTRIBUTORS file
+
+### Changed
+- Changed default matrix type from Dprime to NPMI
+- Use base-pair instead of read-count coverage for window calling
+- Use Orphan windows algorithm to call positive windows
 
 ## [1.1.1] - 2019-02-08
 ### Added